feat: 디렉터리 구조 변경
전체적인 디렉터리 구조를 변경
