+++ src/components/case/CaseListPage.tsx
@@ -9,7 +9,7 @@ import { LawsuitData, LawsuitStatus } from "../../mock/lawsuit/lawsuitTable";
 import caseIdState from "../../states/case/CaseIdState";
 import clientIdState from "../../states/client/ClientIdState";
 import ClientInfo from "../client/ClientInfo.tsx";
-import CaseTable from "./CaseTable.tsx";
+import CaseListTable from "../common/CaseListTable.tsx";
 
 function CaseListPage() {
   const clientId = useRecoilValue(clientIdState);
@@ -91,7 +91,7 @@ function CaseListPage() {
             }}
           />
         </Stack>
-        <CaseTable
+        <CaseListTable
           cases={filteredCases.map((item) => ({
             ...item,
             onClick: () => {
