+++ src/components/common/CaseListTable.tsx
@@ -11,7 +11,7 @@ type Props = {
   cases: (LawsuitData & { onClick: () => void })[];
 };
 
-function CaseTable({ cases }: Props) {
+function CaseListTable({ cases }: Props) {
   return (
     <TableContainer component={Paper}>
       <Table sx={{ minWidth: 650 }} aria-label="simple table">
@@ -52,4 +52,4 @@ function CaseTable({ cases }: Props) {
   );
 }
 
-export default CaseTable;
+export default CaseListTable;
