+++ src/components/join/JoinPage.tsx
@@ -9,7 +9,7 @@ import TextField from "@mui/material/TextField";
 import { useState } from "react";
 import { useNavigate } from "react-router-dom";
 import PopUp from "../common/PopUp";
-import Logo from "../layout/Logo";
+import Logo from "../common/Logo.tsx";
 
 function JoinPage() {
   const [name, setName] = useState("");
