+++ src/components/layout/Layout.tsx
@@ -6,9 +6,9 @@ import userState from "../../states/common/UserState";
 import clientRegisterPopUpState from "../../states/layout/ClientRegisterPopUpOpenState";
 import ClientRegisterPopUp from "../client/ClientRegisterPopUp.tsx";
 import Debug from "./Debug";
-import Header from "./Header";
-import Main from "./Main";
-import SideNavigationBar from "./SideNavigationBar";
+import Header from "./header/Header.tsx";
+import Main from "./main/Main.tsx";
+import SideNavigationBar from "./snb/SideNavigationBar.tsx";
 import ClientRemovePopUp from "../client/ClientRemovePopUp.tsx";
 import clientRemovePopUpOpenState from "../../states/client/ClientRemovePopUpOpenState.tsx";
