+++ src/states/layout/MainNavigationBarState.tsx
@@ -2,7 +2,7 @@ import BadgeIcon from "@mui/icons-material/Badge";
 import BalanceIcon from "@mui/icons-material/Balance";
 import PersonIcon from "@mui/icons-material/Person";
 import { atom } from "recoil";
-import { MainNavigationBarItemState } from "../../components/layout/MainNavigationBarItem";
+import { MainNavigationBarItemState } from "../../components/layout/snb/MainNavigationBarItem.tsx";
 
 export type MainNavigationBarState = {
   curId: number;
